Initial implementation of router
Using react-router-dom to recreate navbar from original startpage
This initial attempt was based on https://medium.com/@krithix/multi-page-website-with-react-in-2020-4dc01267757b
