+++ src/components/Contact.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+class Contact extends Component {
+    render() {
+      return (
+        <div className="Contact">
+          This is the contact page.
+        </div>
+      )
+    }
+  }export default Contact
