+++ src/components/Navbar.css
@@ -10,6 +10,9 @@ The navigation bar is meant to only house links that take the user away
 	position: relative;
     overflow: hidden;
 	margin-bottom: 15px;
+	list-style-type: none;
+	margin: 0;
+	padding: 0;
 }
 /* Style the links inside the navigation bar */
 .Navbar a {
